requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antiberty==0.1.3
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
biopandas>=0.5.1
biopython==1.85
bioservices==1.11.2
biovec==0.2.7
cattrs==23.2.3
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
deepdiff==7.0.1
dnspython==2.6.1
docker==7.0.0
docker-pycreds==0.4.0
easydev==0.13.1
einops==0.7.0
exceptiongroup==1.2.1
fair-esm==2.0.0
foldcomp==0.0.7
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.0.1
gensim==3.4.0
gevent==24.2.1
gitdb==4.0.11
gitpython==3.1.43
graphein[extra]==1.7.7
greenlet==3.0.3
grequests==0.7.0
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
igfold==0.4.0
jaxtyping==0.2.28
joblib==1.4.0
kiwisolver==1.4.5
lightning==2.4.0
lightning-flash==0.8.2
lightning-utilities==0.11.9
line-profiler==4.1.2
loguru==0.7.2
looseversion==1.1.2
lxml==5.2.1
markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
mmtf-python==1.1.3
mpl-chord-diagram==0.3.2
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy<2.0.0
omegaconf==2.3.0
ordered-set==4.1.0
packaging==24.0
pandas==1.5.3
pexpect==4.9.0
platformdirs==4.2.0
plotly==5.21.0
propy3==1.1.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyaaisc==1.2
pydantic==1.10.15
pyfasta==0.5.2
pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
pyprind==2.11.3
pyre-extensions==0.0.30
python-dotenv==1.1.0
pytorch-lightning==1.8.6
rdkit==2023.9.5
regex==2024.4.16
requests-cache==1.2.0
rich==13.7.1
rich-click==1.7.4
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selfies==2.1.1
sentry-sdk==1.45.0
setproctitle==1.3.3
smart-open==7.0.4
smilite==2.3.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
suds-community==1.1.2
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
threadpoolctl==3.4.0
tokenizers==0.19.1
torch-geometric==2.5.3
torcheval==0.0.6
torchmetrics==0.10.3
torchtnt==0.2.3
tqdm==4.66.2
transformers==4.40.0
typeguard==2.13.3
typing-extensions==4.5.0
typing-inspect==0.9.0
url-normalize==1.4.3
wandb==0.16.6
werkzeug==3.0.2
wget==3.2
wrapt==1.16.0
xarray==2024.3.0
xmltodict==0.13.0
yarl==1.9.4
zope-event==5.0
zope-interface==6.3

# PyTorch Geometric extensions (CUDA 12.1 compatible)
pyg-lib @ https://data.pyg.org/whl/torch-2.1.0%2Bcu121/pyg_lib-0.4.0%2Bpt21cu121-cp310-cp310-linux_x86_64.whl
torch-scatter @ https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_scatter-2.1.2%2Bpt21cu121-cp310-cp310-linux_x86_64.whl
torch-sparse @ https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_sparse-0.6.18%2Bpt21cu121-cp310-cp310-linux_x86_64.whl
torch-cluster @ https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_cluster-1.6.3%2Bpt21cu121-cp310-cp310-linux_x86_64.whl
torch-spline-conv @ https://data.pyg.org/whl/torch-2.1.0%2Bcu121/torch_spline_conv-1.2.2%2Bpt21cu121-cp310-cp310-linux_x86_64.whl